lots of pc work, implemented ADC
Did enough work on the status register and memory reading to implement
all of ADC. I'm starting to wonder about a few of my design choices. I'm
having to repeate myself a lot on matching, but I expected this. I
imagine if this ever gets refactored I'll just throw everything in the
master match statement